Redesign asset listing and validation infrastructure
This change builds on initial work in prior commits to fundamentally
redesign the way we manage assets in Bisq.

Previously, assets were "listed" by adding them in several places,
mainly AltCoinAddressValidator and CurrencyUtil. This led to large and
growing classes that do everything and that are prone to merge conflicts
any time there is more than one open pull request to list an asset.

Now we have a new, top-level 'bisq.asset' package (we may want to
extract this to its own repository for several reasons). Within, we have
a first class concept of an 'Asset' (think "crypto asset", as in the most
general term possible for all different kinds of crypto currencies,
tokens, colored coins, etc.). Assets implementations are listed in the
META-INF/services/bisq.asset.Asset service-provider file.

Two subtypes of Asset exist: Coin and Token. 'Coin' is what we have
always meant when saying "altcoin" so far: i.e. a coin that has its own
blockchain. A 'Token' is an asset that does not have its own blockchain,
but rides on top of another through a smart contract, e.g. ERC-20
tokens. And indeed, there is an abstract Erc20Token subtype that captures
this common suitation and provides various conveniences.

In this commit, several assets have been extracted, including Bitcoin,
Bitcoin Cash, Ether, Instacash and Ellaism. Each represents an at least
somewhat different kind of asset, and therefore provided a good basis
for ensuring that the changes in this commit were sufficiently
generalized to accommodate the other asset extractions that will follow
in subsequent commits:

 - Bitcoin is a base currency for Bisq
 - Bitcoin Cash is a typcial Bitcoin-based altcoin
 - Instacash is another Bitcoin-based altcoin
 - Ether is a non Bitcoin-based altcoin
 - Ellaism is an ERC-20 token

A number of changes remain to be made in the new types introduced
here, including Javadoc, and this will be added prior to merging the
series of commits that will make up the overall pull request.

Also note that the tests in AltCoinAddressValidatorTest have been left
intact for now, as a means of double-checking that everything works as
expected from the application level.

It should also be noted that AltCoinAddressValidator is essentially
becoming a translation layer (an "anti-corruption layer" in DDD terms)
where the new 'bisq.asset' world meets the old AltCoinAddressValidator /
CurrencyUtil world that bisq-desktop already knows. It is intentional that
this refactoring has been isolated in this way, leaving the old
"contract" for AltCoinAddressValidator and CurrencyUtil intact. We can
see about refactoring bisq-desktop to deal more directly with the new
AssetRegistry interface in the future, but it is not necessary, and
indeed not desirable to do all of that now in this PR. The idea is to
redesign the underlying infrastructure now, enough to make it more
manageable to deal with going forward.