Lower minimum withdrawal fee to 1 satoshi per byte
Previously, the minimum withdrawal fee was 5 sat/byte, but since the
mempool has been empty for months now, it is no longer necessary to
impose a lower bound at all, because 1 sat/byte transactions clear
easily, usually in one block.

We may re-institute a higher minimum in a future version if and when
the mempool becomes consistently backlogged again. But for now, there's
no sense in forcing users to pay more than is necessary to withdraw
coins from their Bisq wallets.