Prune TradeStatisticsManager#printAllCurrencyStats
This appears to be dead code, see @blabno's inquiry about it at
https://bisq.slack.com/archives/C1267HFD1/p1521801524000052.

If this is in fact *not* dead code, i.e. it gets used to print out
certain reports at release boundaries, I suggest we still kill it here,
and re-write that code in such a way that it (a) uses the new bisq.asset
infrastructure and (b) doesn't require duplicating ticker symbols in
this class.

This work was intended for inclusion in bisq-network/bisq-core#32.