Add supportedCapabilities to Peer
If the AckMessage is sent as a mailbox message we need to know if the
peer supports the capability. If there is no connection already created
or if the connection is new and the capabilities not set by the peer, we
would not know his capabilities. One solution would be to send first a
request fro his capabilities but that adds time delay and complexity.
We use the persisted peer data to store the capability. In the use case
of a trade we have been earlier in contact with the peer so we should
have his capability received. The peer list gets cleaned up with old
entries after a while but we expect that the trade messages are sent
earlier before a peer gets removed from the peers list. And even if we
would not have the peer there the worst thing is that the AckMessage
will not be sent. We log a warning for such cases.

We use a weakReference for the listener for notifying the peers if their
capabilities property gets changed. Peers might get removed from the
list anytime. With the weak reference we don't need to worry about
removing the listeners to avoid memory leaks.