Refine trade feedback message
Previously the message specified that the survey was "one minute" in
duration. This was based on my reading on the statistic on the survey
itself of how long the average user had been taking to fill out the
survey. I realize now that this is not necessarily an accurate
representation of how long it would take to actually fill out the full
form, but rather how long people had been taking before clicking submit.

Now, it simply says that it is a 'short' survey without specifying any
particular duration.