Refactor mutable data handling in State
The data handling with flat maps for each type of mutable data as we
used before feels very un-intuitive to use.
This commit changes the flat map to a tree structure.
We use a map of mutableTx objects. The mutableTx contains the
immutable Tx and a map of mutableTxOutputs as well as mutable fields
like type. The mutableTxOutput contains an immutable TxOutput and
mutable fields like type.

That commit might have broken something in the bonding. Should be fixed
in updates...