Tweak arguments passed to GenesisTxValidator.getGenesisTx()
The genesisTotalSupply argument is added, which earlier was fetched implicitly from
BsqState.

The blockHeight argument is removed, since it refers to the blockheight of the RawTx
argument, which also can be fetched from that object with .getBlockHeight().

Calling code in BsqParser.maybeAddGenesisTx() is updated to respect new call signature.
Since .maybeAddGenesisTx() does not need blockHeight anymore either, it is also dropped here
as well as in the callers in LiteNodeParser and FullNodeParser.