Add adjustment fro multiple ofg 10 EUR for HalCash
With ATMs one can withdraw only multiples of 10 EUR.
We adjust the input values the way that the EUR amount leads to
multiple of 10. MinAmount and amount will be adjusted.

Please note that a similar approach like implemented here for HalCash
can be used for removing the decimal places from fiat amounts to
improve privacy. We keep that for a follow up PR to not mix 2 different
use cases.