Remove iphone 6 warning
Background App refresh need to be enabled (wlan+mobile data) and bisq
enabled (settings/general/Background App refresh). If that is not
enabled the app does not receive the messages even if it is open.
Apple has some weird logic who the name such "features" - you need
Background App refresh enabled to receive data if the app is not in
background.... I think another reason to push Apple in background...