Add basic OfferUtilTest unit tests
They verify that OfferUtil.getAdjustedAmount() behave as we expect
as a starting point for refactors or simplification of this fairly
tricky method.

As the TODO message for the last test case describes, it is not
clear to me that the specific input values should return any value
at all, since they seem to represent constraints which cannot all
be met at the same time. We may want to change the behavior under
such circumstances (i.e by raising an exception) in the future.