Let onTimeout call on onSuccess instead of onFailure
- In case the tx broadcast did not completed in time we call the
onTimeout handler which logs a waring but optimistically assume that
the broadcast will still succeed later and call onSuccess.
This strategy has advantages over calling onFailure as it might be more
often the case that the broadcast was just slow but successful later
then that the broadcast fails at all.