Refactor OpReturnParser for VoteReveal
Add isPhaseValid method.

We try to get rid of the sub parser classes for opReturn and remove the
checks which are not directly related to the OpReturn data like the fee
and the phase checks. We set the verifiedOpReturnType in the
TxOutputParser and do the check for fee and phase in the TxParser.