Refactor OpReturnParser and TxOutputParser methods
OpReturnParser.parseAndValidate() now takes TempTxOutput as argument,
and no longer needs TempTx or index. It takes a boolean lastOutput
instead of nonZeroOutput, as the latter can be deduced from the
TempTxOutput.

The OpReturnParser.validate() method now takes blockHeight, and
no longer needs the TempTx, which was only used for logging an
error case.

For TxOutputParser.processTxOutput(), it now takes a boolean lastOutput
argument, which is passed on to OpReturnParser.parseAndValidate() as
mentioned above. It no longer needs the TempTx either.

Also add docstring for TxOutputParser.processTxOutput().

Use TxOutput instead of Tx.
Remove unneeded arguments.