Remove TempTx argument from TxParser.getOptionalOpReturnType()
The only use of this argument was to log different messages in case
of unexpected transactions, and we likely can produce better debugging /
testing info in other ways if this is needed. It seems better to pass
in only the arguments which we need.

Also add docstring to the method.